openhcl: misc guest_arch cleanups
