--- openhcl/diag_server/build.rs
@@ -1,8 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-#![expect(missing_docs)]
-
-fn main() {
-    build_rs_guest_arch::emit_guest_arch()
-}
