+++ openhcl/underhill_core/src/vp.rs
@@ -126,6 +126,7 @@ impl VpSpawner {
         saved_state: Option<vmcore::save_restore::SavedStateBlob>,
         control: Option<&mut IdleControl>,
     ) -> Option<vmcore::save_restore::SavedStateBlob> {
+        #[allow(unreachable_patterns)]
         let r = match self.isolation {
             virt::IsolationType::None | virt::IsolationType::Vbs => {
                 self.run_backed_vp::<virt_mshv_vtl::HypervisorBacked>(saved_state, control)
@@ -141,7 +142,6 @@ impl VpSpawner {
                 self.run_backed_vp::<virt_mshv_vtl::TdxBacked>(saved_state, control)
                     .await
             }
-            #[cfg(guest_arch = "aarch64")]
             _ => unimplemented!(),
         };
         match r {
