+++ vmm_core/src/emuplat/mod.rs
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+pub mod apic;
 pub mod gic;
 pub mod hcl_compat_uefi_nvram_storage;
 pub mod ioapic;
